pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inboxpay-fastapi"
version = "0.1.0"
description = "FastAPI application for inbox-based payment processing"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "openai",
    "jinja2",
    "agentmail",
    "google-generativeai",
]

[project.optional-dependencies]
dev = [
    "types-requests",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
